Refactor models command formatters into dedicated modules
