+++ mesh-llm/src/models/mod.rs
@@ -22,9 +22,11 @@ pub use maintenance::{run_update, warn_about_updates_for_paths};
 pub use resolve::{
     download_exact_ref, find_catalog_model_exact, installed_model_capabilities,
     installed_model_display_name, resolve_model_spec, show_exact_model,
-    show_model_variants_with_progress, ShowVariantsProgress,
+    show_model_variants_with_progress, ModelDetails, ShowVariantsProgress,
 };
-pub use search::{search_catalog_models, search_huggingface, SearchArtifactFilter, SearchProgress};
+pub use search::{
+    search_catalog_models, search_huggingface, SearchArtifactFilter, SearchHit, SearchProgress,
+};
 pub use topology::{infer_local_model_topology, ModelMoeInfo, ModelTopology};
 
 fn build_hf_api(progress: bool) -> Result<Api> {
